requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
python-dotenv
alembic
aiosmtplib
httpx
fastapi-mail
python-multipart
python-jose[cryptography]
psycopg[binary]>=3.1 
passlib[bcrypt]
psutil
PyPDF2
openpyxl
python-docx
